// next/lib/api/types.ts
export interface PublicKey {
  id: number;
  userId: number;
  hash: string;
  data: string;
  isRootKey: boolean;
  enabled: Date | null;
  enabledBy: string | null;
  createdAt: Date;
  updatedAt: Date;
}
// ...
export interface UserResponse {
  /**
    * The database id of the user
    * @isInt
    */
  id: number;
  /**
    * The unique user sub
    */
  sub: string;
  /**
    * The name of the user
    */
  name: string;
  /**
    * The email of the user
    */
  email: string;
  /**
    * The date the user was created
    */
  createdAt: Date;
  /**
    * The date the user was last updated
    */
  updatedAt: Date;
  /**
    * The public keys of the user
    */
  keys: PublicKey[];
}